a small restructure
