+++ src/main/java/oripa/gui/presenter/main/CircleCopyDialogPresenter.java
@@ -21,8 +21,8 @@ package oripa.gui.presenter.main;
 import oripa.domain.paint.CircleCopyParameter;
 import oripa.domain.paint.PaintContext;
 import oripa.domain.paint.circlecopy.CircleCopyCommand;
+import oripa.gui.view.ViewScreenUpdater;
 import oripa.gui.view.main.CircleCopyDialogView;
-import oripa.gui.view.main.ViewScreenUpdater;
 import oripa.util.Command;
 
 /**
