+++ src/main/java/oripa/gui/view/main/ViewUpdateSupport.java
@@ -18,6 +18,8 @@
  */
 package oripa.gui.view.main;
 
+import oripa.gui.view.ViewScreenUpdater;
+
 /**
  * @author OUCHI Koji
  *
